Allow to fix diffusion coefficients in the model
In case those parameters have already been estimated from an independent source.

To use it, simply set the lower bound and the upper bound of the corresponding
diffusion coefficient to the independently estimated value.

**Implementation details:** Look at the difference between LB and UB, if
under a float threshold, consider that this parameter should be fixed, and
declare it as is.

**API change:** this should allow not to break the API, and provide natural
evolution for aplications that depend on fastspt, such as Spot-On.